MISSIONS: make octagon mission generic for any go to pinger
