+++ command/sub8_missions/missions/autonomous.py
@@ -7,8 +7,8 @@ from mil_misc_tools import text_effects
 import start_gate
 import buoy
 import align_path_marker
-import navigation_gate
-import surface_octagon
+import go_to_pinger
+import surface
 
 
 fprint = text_effects.FprintFactory(title="AUTO_MISSION").fprint
@@ -26,7 +26,8 @@ def do_mission(sub):
         yield buoy.run(sub)
         yield align_path_marker.run(sub)
         yield navigation_gate.run(sub)
-        yield surface_octagon.run(sub)
+        yield go_to_pinger.run(sub)
+        yield surface.run(sub)
     except Exception as e:
         fprint("Error in Chain 1 missions!", msg_color="red")
         print e
