+++ command/sub8_missions/missions/surface.py
@@ -4,5 +4,5 @@ from txros import util
 @util.cancellableInlineCallbacks
 def run(sub_singleton):
     print "Surfacing"
-    yield sub_singleton.move.depth(0.6).go()
+    yield sub_singleton.move.depth(0.1).go()
     print "Done surfacing"
